i#7644: Add invariant checks for template availability (#7677)
Adds drmemtrace invariant checks to ensure that the syscall template
file has a default template.

Adds drmemtrace invariant checks to ensure that the switch template file
has a template for both kinds of context switches.

The core-sharded invariant checks added in #7665 require all instances
of syscalls and switches to be injected. So this property should also be
verified in the template files themselves, or else we'll know about the
missing kernel traces late after injection.

Relaxes the chunk instr count check for kernel trace template files, and
updates the mock switch template (which is a zip file with chunks) to
remove the chunk_instr_count marker.

Issue: #7644, #6684