i#7676: Refactor aarchxx private TLS to match the actual layout (#7679)
The aarchxx private library TLS was laid out like x86 with the
app/library TLS before the TCB, but in actuality it needs to look more
like RISC-V with the TCB before the app/library TLS.

Re-orders and updates the offset from the thread pointer.
Updates the ELF_R_TLS_DESC relocation and other places where the
direction is specified.
Updates the static client thread TLS page to use 2 pages so the thread
pointer can point to the middle, allowing for a negative offset.

Leaves some space for future pthread struct increases.
Adds decoding of pthread_mutex_consistent() to set the pthread tid field
for platforms different than the 2.38 default.

Keeps 32-bit arm with the checked-in pthread layout, as that came from
aarch32.
Removes the TCB padding to correctly work with the new ordering.

Issue: #1569, #7670, #7676
Fixes #7676