i#7685: Initial implementation of the BBV client
Initial implementation of client that generates Basic Block Vectors
(BBVs), which capture Basic Blocks (BBs) execution frequencies within
a user-provided instruction interval.

The client registers two events:
- `drmgr_register_bb_instrumentation_event()` to instrument BBs.
- `drmgr_register_exit_event()` to save the collected BBV data to a file.

Example usage:
```
./bin64/drrun -c api/bin/libbbv.so -instr_interval 10000 -- ls .
```
Generates a proc.APP_NAME.PID.UNIQUE_ID.bb file.

The instrumentation is a single clean call to a function.
Because clean calls are expensive we plan to inline the counter updates
in the instrumented BBs themselves and jump to the rest of the
instrumentation as a cleaan call only when the given instruction
interval is reached.

Currently we generate a unique ID for BBs using the BB first instruction
PC (its absolute value). This is not a reliable method. We plan to use
the `drmodtrack` library to compute instruction module offsets instead.

Adds a new `hashtable_t` iterator:
`hashtable_apply_to_all_key_payload_pairs_user_data()` that exposes the
key of each payload when iterating over the elements of an
`hashtable_t`. This is necessary as we need to save both the BB ID (key)
and hit count (payload) once we reach the end of an instruction
interval. The alternative would be to add the key to the payload, but
it seems unnecessary and consumes more memory.

Issue #7685