i#5426: Mark instrs in AArch64 mangling epilogue and emulate them.
In mangling instrs that use the stolen register, mark the instrs
that restore the register state as mangling epilogue.

In recreate_app_state_from_ilist emulate those instructions to
advance the app state to after the app instr.

This fixes #5426, for which a test is added, but there are probably
other translation failures on AArch64, and similar failures on other
architectures.

The test uses a new function, adaptive_retry, in tools.c, which
is abstracted from signal_pre_syscall.c.

Issue: #3307, #3329, #5426.
Fixes #5426

Change-Id: I825dfbe7188625777abf7b2f26e7170ec7ed2601