i#6684: Skip chunk invariants for dynamic core-sharded traces
Skips drmemtrace invariant checks related to chunk instr counts and chunk ordinals when processing traces that are dynamically core-shareded. Such traces are not written down in the zip format, therefore are not expected to have the same notion of chunks.

Adds unit tests that verify these relaxations are applied for core-sharded traces but not core-sharded-on-disk traces.

Issue: #6684