i#6684: Skip chunk invariants for dynamic core-sharded traces (#7689)
Skips drmemtrace invariant checks related to chunk instr counts and
chunk ordinals when processing traces that are dynamically core-sharded.
Such traces are not read from disk at all, therefore are not expected to
have the same notion of chunks.

Adds unit tests that verify that these relaxations are applied for
dynamically core-sharded traces but not core-sharded-on-disk traces.

Issue: #6684