i#7576 AArch64 detach: Fix intermittent hang
After porting the api.detach_state test to AArch64 we started to notice
an intermittent failure caused by the test hanging when trying to detach
while the check thread is performing a code cache flush.

The test has two threads. Thread 1 is performing the detach and thread 2
is performing the code cache flush. Both threads call
synch_with_all_threads() and if they do it at the same time with thread
1 taking the synch_lock first  we get a deadlock. Thread 1 is waiting
for thread 2 to suspend in a spot with a valid mcontext, thread 2 is
waiting for thread 1 to give up the synch_lock and is suspended without
a valid mcontext so neither thread can progress.

I have fixed this by changing the AArch64 cache flush handling code to
call synch_with_all_threads() with a valid mcontext so the detach thread
will no longer get stuck.

This required adding an extra argument to
flush_fragments_in_region_start() to specify the synch permissions and
plumbing it through to the synch_with_all_threads() call. All other
callers of this function (and the ones it wraps) have been updated to
set THREAD_SYNCH_NO_LOCKS_NO_XFER permission which is the value that
was previously used in flush_fragments_synchall_start().

This also reverts commit 29120ddcdca486aec70fadf87b410409d548e6e3 so
failures of this test will no longer be ignored.

Issue: #7576