i#7046 memory dump: Add a vvar section to the memory dump file. (#7699)
Add an empty section for vvar to memory dump file. 

Code in vdso section refers to addresses in vvar section which is being
dropped in the current implementation.
The section is not readable, so a program header with the actual memory
size and zero file size is added.

Modify suite/tests/client-interface/memory_dump_test.templatex to look
for vvar and vdso section in the memory dump file.

Issue: #7046