i#7701: Merge split-asm atomic operations (#7702)
Unifies the split atomic operations on x86 and aarch32 where the result of an
operation was computed from the condition codes via a separate asm
statement. The previous separate asm statements allowed the compiler to insert
code in between that clobbers the flags, which we actually observed in the
inlined d_r_mutex_trylock inside try_recursive_lock. We fix that by using a
single unified asm statement.

New atomic macros are created that perform the flags transfer into a
result variable for all affected operations:
+ atomic_inc_and_test()
+ atomic_dec_and_test()
+ atomic_dec_becomes_zero()
+ atomic_compare_exchange_int()

Changes atomic_dec_and_test()'s documentation to only return
whether the result is negative, as that's what x86 and aarch32
were doing, despite the docs. Changes the Windows, aarch64,
and RISCV code to match this. This is only used for mutexes
who have only one valid negative value so this is safe.

We already have some unit tests in unit_test_atomic_ops() in arch.c for
basic regression sanity checks.

Tested with the internal build and confirmed there is no longer an
instruction in between cmpxchg and setcc.

Fixes #7701