i#7487: fix tool's crashes due to cfg thunk implemented on 24h2 and above
This commit fixes the crashes on the tool when it encounters cfg thunks `guard_dispatch_icall$thunk$xxx` that was dereferenced to extended image memory located at ntdll by OS kernel, which leads to invalid memory for private library.

Apart from that, `redirect_GetModuleFileNameW` seems to be broken as it returns the ascii version for the filename instead of unicode version.

Fixes #7487