i#7685 DrPoints: initial implementation of the BBV client (#7687)
Initial implementation of DrPoints, a client that generates Basic Block
Vectors (BBVs), which capture Basic Blocks (BBs) execution frequencies
within a user-provided instruction interval.

The client registers three events:
- `drmgr_register_bb_instrumentation_event()` to instrument BBs.
- `drmgr_register_exit_event()` to save the collected BBV data to a
file.
- `drmgr_register_thread_init_event()` to check for multi-threaded
programs and abort if so (multi-threaded program are not currently
supported).

Example usage:
```
drrun -t drpoints -instr_interval 10000 -- ls .
```
Generates a `drpoints.APP_NAME.PID.UNIQUE_ID.bbv` file.

This is an initial, working implementation, but there are missing
pieces:

1) Instrumentation optimizations for AARCH64 and 32 bit architectures.
The instrumentation is a single clean call to a function in this case.
Because clean calls are expensive, for X86_64, we inline the counter
updates in the instrumented BBs themselves and jump to the rest of the
instrumentation as a clean call only when the given instruction interval is
reached. We plan to do the same for, at least, AARCH64 architectures.

2) Use of absolute PC values.
Currently we generate a unique ID for BBs using the BB first instruction
PC (its absolute value). This is not a reliable method. We plan to use the
`drmodtrack` library to compute instruction module offsets instead.

3) Add documentation.
Generate .dox documentation.

Issue #7685