i#7729: Fix incorrect drmemtrace addr for stolen reg
Fixes a bug where drmemtrace records the wrong address for a load with
no index and a base of the stolen register.

Adds a test case to burst_aarch64_sys which fails without the fix.

Adds a drutil base reg test case to the stolen-reg-index test as well,
though drutil operates correctly here.

Fixes #7729