i#7685 DrPoints: use drmodtrack to get BB id (#7733)
Improves BB identification by using module-relative offsets instead of
absolute PCs, making it robust against dynamic library loading.
It uses the drmodtrack library to get a unique <module_index, offset>
pair for each BB. Then, to create a scalar hash from this pair, it uses
Szudzik's elegant pairing function. This function maps the
two-dimensional <module_index,offset> pair to a single, unique number,
which is then used as the key in the hashtable that contains the BB id.
This avoids the complexity of a nested hash table.

Issue #7685