i#7718: Fix compile errors with Clang 18+ (#7719)
Drcachesim uses the C++ utility function offsetof on non-POD structs.
This results in a warning with Clang 18+. Remove the use of offsetof
for binary compatibility.

Fix a couple of other compiler errors so that DR can be built with
Clang 18+. Further changes are required in order for the tests to build.

Fixes: #7718