i#7722: Move AArchXX TP to base of TCB head (#7738)
The thread pointer (TP, the value in the tpid* register) for glibc on
AArchXX was pointing to the end of the TCB head after PR #7679 fixed the
TLS layout to have the TLS correctly live after the TCB. However, unlike
RISC-V, AArchXX has the TP point to the base of the TCB head.

Moving the TP revealed a bug in the R_AARCH64_TLS_TPREL relocation where
the sign is inverted, which is also fixed here (it should have been
fixed in PR #7679 when it fixed R_AARCH64_TLSDESC). This manifested in
the client.tls test as a crash in glibc init.

Issue: #7722, #7676
Fixes #7722