i#7685 DrPoints: add inline counter update for AARCH64 (#7737)
Introduces the inline counter update optimization for AARCH64 in the
DrPoints client. Previously, this optimization was only implemented for
X86_64. It uses the `tbz` (test bit and branch if zero) instruction to
check (w/o aflags) if the instruction counter has become negative by
testing the sign bit and branching to a clean call that saves the basic
block vector if the check succeeds. Because `tbz` does not branch
to `save_bbv()` when the instruction counter reaches 0 (it branches
only when it's negative), the initial instruction count is decremented
by 1 in this case to keep the same behavior:
"branch when instr_count <= 0".

Adds a new assembly test case specifically for AARCH64, which
mirrors the X86_64 test.

Issue #7685