i#7588 drsyscall: fix a boundary condition bug in drsyscall_iterate_records. (#7736)
There is a bug in drsyscall_iterate_records() where it stops to process
syscall record if a record ends exactly at the end of its buffer. This
bug affects both drsyscall_record_trimmer and drsyscall_record_viewer.

The root cause of this failure is not in the syscall record collection
but rather in the iteration logic used by drsyscall_iterate_records()
during subsequent processing.

Add a unit test which has a syscall record which ends exactly at the end
of the buffer. Without the fix, the syscall record
DRSYS_RECORD_END_TIMESTAMP will not be read.

Issue: #7588