i#7685 DrPoints: BB ID bug fix (#7755)
The key to `bb_id_table` is not just an integer that we can store into a `void*`.
It's a struct that needs to be allocated when adding a new entry into the table.
Previously we allocated the key on the stack, hence losing it after the BB
instrumentation function. We now allocate it on the heap with `dr_global_alloc()`.
We also configure the hashtable setting a function to free its keys when deleted.

Issue #7685